+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modules for easily enabling/disabling groups of methods.
 - Rich text support.
 
+## [0.1.2] - 2021-04-10
+## Changes
+- "help" command changed to "command" and it now lists either all commands or filtered by substring
+- to list commands by module, use "module-commands"
+## Fixes
+- Fixed editor Consoled Window losing focus on submit
+
 ## [0.1.1] - 2021-04-09
 ## Changes
 - Value adapters now support advanced type adapting using the bool CanAdapt(Type targetType) method.